* Migrate some of the DOS code into msdos/ * Move the packet driver class definitions into iface.h; they're used   as interface type flags too (for now?) * Don't build the arcdump.c file; it's not used internally.
These files, with the exception of arcdump, aren't actually even
built in this port.
