+++ msdos/dirutil.c
