+++ msdos/dma.c
