+++ msdos/gopint.c
