+++ msdos/gormcb.c
