+++ msdos/pc.c
